Push the protocol user to send the channel info once connected Add bitrate function and use enums in the audiotransportdetails Provide a default for iOS

+++ Sources/JamulusProtocol/Models/ApiConsts.swift
@@ -2,7 +2,7 @@
 import Foundation
 
 public enum ApiConsts {
-  public static let sampleRate48kHz: UInt32 = 48000
+  public static let sampleRate48kHz: Float = 48000
   public static let frameSamples64: UInt32 = 64
   public static let defaultPort: UInt16 = 22124
   public static let packetHeaderSize = 9
